+++ src/pages/Checkout/Checkout.js
@@ -1,55 +1,15 @@
-import React, { useState } from "react";
+import React from "react";
 
 import CheckoutProfile from "./CheckoutProfile";
+import FormContextProvider from "../../components/ContextComponents/FormContextProvider";
 
 import { PROFILE } from "../../utils/constants";
 
-import FormContext from "../../context/form-context";
-
-const initialValueFormData = {
-  name: "",
-  email: "",
-  phone: "",
-  address: "",
-  city: "",
-  postCode: "",
-  country: "",
-  card: {
-    cardName: "",
-    cardNumber: "",
-    cardDate: "",
-    cardCVV: "",
-    cardTerms: false,
-  },
-  payPal: {
-    payPalUser: "",
-    payPalPassword: "",
-  },
-  applePay: {
-    applePayUser: "",
-    applePayPassword: "",
-  },
-};
-
 function Checkout({ processStep }) {
-  const [formData, setFormData] = useState(initialValueFormData);
-
-  function updateFormData(newData) {
-    setFormData((prevData) => {
-      console.log({ ...prevData, ...newData });
-      return { ...prevData, ...newData };
-    });
-  }
-
   return (
-    <FormContext.Provider
-      value={{
-        data: formData,
-        setData: updateFormData,
-      }}
-    >
+    <FormContextProvider>
       {processStep === PROFILE && <CheckoutProfile />}
-    </FormContext.Provider>
+    </FormContextProvider>
   );
 }
